+++ crates/orchestrator/src/queue/sqs/mod.rs
@@ -1,14 +1,28 @@
+use std::collections::HashMap;
 use std::time::Duration;
 
-use crate::queue::job_queue::JOB_PROCESSING_QUEUE;
+use crate::queue::job_queue::{
+    JOB_HANDLE_FAILURE_QUEUE, JOB_PROCESSING_QUEUE, JOB_VERIFICATION_QUEUE, WORKER_TRIGGER_QUEUE,
+};
 use async_trait::async_trait;
 use color_eyre::Result;
+use lazy_static::lazy_static;
 use omniqueue::backends::{SqsBackend, SqsConfig, SqsConsumer, SqsProducer};
 use omniqueue::{Delivery, QueueError};
 use utils::env_utils::get_env_var_or_panic;
 
 use crate::queue::QueueProvider;
 pub struct SqsQueue;
+
+lazy_static! {
+    /// Maps Queue Name to Env var of queue URL.
+    pub static ref QUEUE_NAME_TO_ENV_VAR_MAPPING: HashMap<&'static str, &'static str> = HashMap::from([
+        (JOB_PROCESSING_QUEUE, "SQS_JOB_PROCESSING_QUEUE_URL"),
+        (JOB_VERIFICATION_QUEUE, "SQS_JOB_VERIFICATION_QUEUE_URL"),
+        (JOB_HANDLE_FAILURE_QUEUE, "SQS_JOB_HANDLE_FAILURE_QUEUE_URL"),
+        (WORKER_TRIGGER_QUEUE, "SQS_WORKER_TRIGGER_QUEUE_URL"),
+    ]);
+}
 
 #[async_trait]
 impl QueueProvider for SqsQueue {
@@ -31,12 +45,11 @@ impl QueueProvider for SqsQueue {
     }
 }
 
+/// To fetch the queue URL from the environment variables
 fn get_queue_url(queue_name: String) -> String {
-    if queue_name == JOB_PROCESSING_QUEUE {
-        get_env_var_or_panic("SQS_JOB_PROCESSING_QUEUE_URL")
-    } else {
-        get_env_var_or_panic("SQS_JOB_VERIFICATION_QUEUE_URL")
-    }
+    get_env_var_or_panic(
+        QUEUE_NAME_TO_ENV_VAR_MAPPING.get(queue_name.as_str()).expect("Not able to get the queue env var name."),
+    )
 }
 
 // TODO: store the producer and consumer in memory to avoid creating a new one every time